voice_agent_backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
dnspython==2.7.0
elevenlabs==2.7.1
eventlet==0.40.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
multidict==6.6.3
narwhals==1.47.1
numpy==2.3.1
openai==1.97.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
plotly==6.2.0
propcache==0.3.2
protobuf==6.31.1
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
streamlit==1.47.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
twilio==9.6.5
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.1
